+++ changelog.md
@@ -2,11 +2,17 @@
 
 Changelog was added at version 2.0.3 for internal reference.
 
+### 2.0.4 | 31/03/2021
+
+###### Rule Changes
+
+- `multiline-ternary` set to `0`
+
 ### 2.0.3 | 14/03/2021
 
 ###### @babel/eslint-parser
 
-  - `parserOptions.requireConfigFile` set to `false`
+- `parserOptions.requireConfigFile` set to `false`
 
 ###### Miscellaneous